+++ admin/js/admin.js
@@ -21,4 +21,17 @@
 		});
 	});
 
+	// Admin What's New dismiss
+	$(document).on('click', '#xen-admin-dismiss-whats-new', function () {
+		var $btn  = $(this);
+		var nonce = $btn.data('nonce');
+		$btn.prop('disabled', true).text('…');
+		$.post(xenAdmin.ajaxUrl, {
+			action: 'xen_admin_dismiss_whats_new',
+			nonce:  nonce,
+		}, function () {
+			$('#xen-admin-whats-new').fadeOut(300, function () { $(this).remove(); });
+		});
+	});
+
 }(jQuery));